config: move to klplib package
