+++ klpbuild/klplib/ibs.py
@@ -23,7 +23,7 @@ from lxml.objectify import SubElement
 from natsort import natsorted
 from osctiny import Osc
 
-from klpbuild.config import Config
+from klpbuild.klplib.config import Config
 from klpbuild.klplib.utils import ARCH, ARCHS, get_all_symbols_from_object, get_elf_object, get_elf_modinfo_entry, get_cs_branch, get_kgraft_branch, filter_codestreams
 
 class IBS(Config):
