+++ klpbuild/klplib/templ.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from mako.lookup import TemplateLookup
 from mako.template import Template
 
-from klpbuild.config import Config
+from klpbuild.klplib.config import Config
 from klpbuild.klplib.utils import ARCHS, fix_mod_string, get_mail
 
 TEMPL_H = """\
