+++ klpbuild/plugins/extractor.py
@@ -21,7 +21,7 @@ from filelock import FileLock
 from natsort import natsorted
 
 from klpbuild.klplib import utils
-from klpbuild.config import Config
+from klpbuild.klplib.config import Config
 from klpbuild.klplib.templ import TemplateGen
